requirements.txt
langchain==0.1.0
langgraph==0.0.20
llama-index==0.9.0
chromadb==0.4.22
sentence-transformers==2.3.1
tavily-python==0.3.0
duckduckgo-search==4.1.0
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
rich==13.7.0
typer==0.9.0

# Optional: Advanced agent frameworks
crewai==0.1.0
pyautogen==0.2.0

# Optional: Enhanced monitoring
prometheus-client==0.19.0
psutil==5.9.7

# Optional: Advanced visualization
plotly==5.18.0
seaborn==0.13.1
